+++ src/main.rs
@@ -31,11 +31,11 @@ fn random_scene() -> World {
 
     for a in -11..=11 {
         for b in -11..=11 {
-            let choose_mat: f64 = rng.gen();
+            let choose_mat: f32 = rng.gen();
             let center = Point3::new(
-                (a as f64) + rng.gen_range(0.0..0.9),
+                (a as f32) + rng.gen_range(0.0..0.9),
                 0.2,
-                (b as f64) + rng.gen_range(0.0..0.9),
+                (b as f32) + rng.gen_range(0.0..0.9),
             );
 
             if choose_mat < 0.8 {
@@ -103,7 +103,7 @@ fn ray_color(r: &Ray, world: &World, depth: u64) -> Color {
         return Color::new(0.0, 0.0, 0.0);
     }
 
-    if let Some(rec) = world.hit(r, 0.001, f64::INFINITY) {
+    if let Some(rec) = world.hit(r, 0.001, f32::INFINITY) {
         if let Some((attenuation, scattered)) = rec.mat.scatter(r, &rec) {
             attenuation * ray_color(&scattered, world, depth - 1)
         } else {
@@ -118,9 +118,9 @@ fn ray_color(r: &Ray, world: &World, depth: u64) -> Color {
 
 fn main() {
     //Imagen por si aca
-    const ASPECT_RATIO: f64 = 16.0 / 9.0;
+    const ASPECT_RATIO: f32 = 16.0 / 9.0;
     const IMAGE_WIDTH: u64 = 500;
-    const IMAGE_HEIGHT: u64 = ((IMAGE_WIDTH as f64) / ASPECT_RATIO) as u64;
+    const IMAGE_HEIGHT: u64 = ((IMAGE_WIDTH as f32) / ASPECT_RATIO) as u64;
     const SAMPLES_PER_PIXEL: u64 = 50;
     const MAX_DEPTH: u64 = 5;
 
@@ -157,11 +157,11 @@ fn main() {
                 let mut pixel_color = Color::new(0.0, 0.0, 0.0);
                 for _ in 0..SAMPLES_PER_PIXEL {
                     let mut rng = rand::thread_rng();
-                    let random_u: f64 = rng.gen();
-                    let random_v: f64 = rng.gen();
+                    let random_u: f32 = rng.gen();
+                    let random_v: f32 = rng.gen();
 
-                    let u = ((i as f64) + random_u) / ((IMAGE_WIDTH - 1) as f64);
-                    let v = ((j as f64) + random_v) / ((IMAGE_HEIGHT - 1) as f64);
+                    let u = ((i as f32) + random_u) / ((IMAGE_WIDTH - 1) as f32);
+                    let v = ((j as f32) + random_v) / ((IMAGE_HEIGHT - 1) as f32);
 
                     let r = cam.get_ray(u, v);
                     pixel_color += ray_color(&r, &world, MAX_DEPTH);
